requirements.txt
google-generativeai>=0.5.0
playwright>=1.40.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pillow>=10.2.0
